requirements.txt
streamlit
flask
pyjwt
requests
pytest
python-dotenv
google-generativeai
flask-cors
pandas
pytest
werkzeug
gtts
SpeechRecognition
langdetect
webdriver-manager
gunicorn


